updated to use Sui --> Fuji repo from WH
